split js code in subfolders
