+++ apps/thunder_phoenix/.gitignore
@@ -8,6 +8,11 @@
 
 # Generated by parcel
 /priv/static/assets/*
+# Genrated by unknown
+/priv/static/js/*
+/priv/static/*.js
+/priv/static/*.css
+/priv/static/*.map
 
 # Generated on crash by the VM
 erl_crash.dump
