+++ apps/thunder_phoenix/lib/thunder_phoenix_web/static/js/app.js
@@ -1,7 +1,22 @@
 import "phoenix_html";
 import "./phoenix";
 
-import "./chat_socket";
-import "./sim_basic_socket";
+if (document.getElementById("chat-input")) {
+  import("./chat/chat_socket").then(function(page) {
+    console.log("chat loaded");
+  });
+}
+
+if (document.getElementById("reverse-input")) {
+  import("./reverse/sim_basic_socket").then(function(page) {
+    console.log("basic loaded");
+  });
+}
+
+if (document.getElementById("counter")) {
+  import("./counter/counter").then(function(page) {
+    console.log("counter loaded");
+  });
+}
 
 import "../css/app.scss";
