+++ apps/thunder_phoenix/lib/thunder_phoenix_web/static/js/chat/chat_socket.js
@@ -1,4 +1,4 @@
-import socket from "./user_socket";
+import socket from "../user_socket";
 
 let channel = socket.channel("room:lobby", {});
 let chatInput = document.querySelector("#chat-input");
