+++ apps/thunder_phoenix/lib/thunder_phoenix_web/static/js/counter/counter.js
@@ -1,0 +1,1 @@
+console.log("counter");
